app/requirements.txt
absl-py==1.4.0
altgraph==0.17.3
asttokens==2.2.1
astunparse==1.6.3
auto-py-to-exe==2.32.0
backcall==0.2.0
beautifulsoup4==4.11.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Eel==0.14.0
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
flatbuffers==23.3.3
future==0.18.2
gast==0.3.3
gdown==4.6.4
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.22.0
importlib-metadata==6.0.0
imutils==0.5.4
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter-client==8.0.3
jupyter-core==5.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mtcnn==0.1.1
nest-asyncio==1.5.6
networkx==2.8.6
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.0
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tifffile==2022.8.12
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
whichcraft==0.6.1
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
